DebianBugTracker: handle get_status() queries in batches
Since jessie, the Apache on bugs.debian.org has some rate limiting
activated, which leads to 500 Internal Server Error when too many bugs
statuses are queried over SOAP at the same time.

This is fixed in python-debianbts since 1.13, 1.14 fixed a regression but
this change is not in stable.

So let's do the same ourselves, with code stolen from python-debianbts.

This change can be reverted once python-debianbts 1.14 is available.